Added new fotos in movies objects.
If applied, this commit will have added new images in each object of movies arrays.
In this case was used a require built-in function to import the pictures. This process was possible because the pictures folder (img) is inside of src folder. Even though to keep the img folder in the public folder should be the future best practice, the require function looked the better for the moment necessities.