Updated with function component LoginView that defines the basic components.
If applied, this commit will have updated the LoginView with a function Component that contains the basic components.
Using Hooks was defined the const that will set a state.
First using the destructure we give to variables the useState() (the state of the fields) first to username(empty string(as we found it in the beginning)) and then setUsername that will updated the state(same for password).

Was added as well the propTypes.