Updated onLoggedIn method, created getMovies method and created onLoggedOut method.
If applied, this commit will have Updated onLoggedIn method, created getMovies method and created onLoggedOut method.
In onLoggedIn method was defined a state and used the localStorage to keep the token and user to remember the browser that the user is still active and doesn't made him to log in again.
The getMovies method was created to avoid repetition on the code since the '/movies' endpoint needed to be used in the moment that we Log and every time the user loads the page.